requirements.txt
pymongo==4.7.0
robotframework==7.0
robotframework-appiumlibrary==2.0.0

